dms_backend/requirements.txt
bcrypt==4.3.0
blinker==1.9.0
click==8.2.0
colorama==0.4.6
Flask==3.1.1
flask-cors==6.0.0
Flask-JWT-Extended==4.7.1
greenlet==3.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql==0.0.3
mysql-connector-python==9.3.0
mysqlclient==2.2.7
PyJWT==2.8.0
python-dotenv==1.0.1
typing_extensions==4.13.2
Werkzeug==3.1.3
